VMM. Minor change.

+++ iaas/virtual-machine-manager/opta-vm-placement/src/main/java/es/bsc/vmplacement/scorecalculators/ScoreCalculatorConsolidation.java
@@ -55,7 +55,7 @@ public final class ScoreCalculatorConsolidation implements SimpleScoreCalculator
     }
 
     private int calculateSoftScore(ClusterState solution) {
-        return (int) -solution.calculateCumulativeUnusedCpuPerc();
+        return -solution.calculateCumulativeUnusedCpuPerc();
     }
 
 }
